requirements.txt
#Seleciona o IA da google
langchain-google-genai==2.0.9

#Core Frameworks
langchain==0.3.25
langchain-community==0.3.25
langgraph==0.4.8

#Utilidades e Vetores
python-dotenv==1.1.0
faiss-cpu==1.11.0
pypdf==5.6.0